Debugging scripts

+++ mcrbin/stop.csh
@@ -1,8 +1,11 @@
 #!/bin/csh
 
-echo "stop.csh" >> /home/blue/2015/log/log.log
-
 set id = `ps -e -u $USER | grep rpoldaemon | awk '{ print $1}'`
-kill -s TERM $id
-echo "rpoldaemon stopped!"
-exit $?
+if ( "$id" != "" ) then
+    kill -s TERM $id
+    echo "rpoldaemon stopped!"
+    exit 0
+else
+    echo "rpoldaemon is not running!"
+    exit 1
+endif
